feat: creation and search for rarity

+++ gamification-api/src/main/java/io/github/thallesyan/gamification_api/infrastructure/web/dto/response/RarityResponseDTO.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 @Data
 public class RarityResponseDTO {
     private RarityEnum value;
-    private Double points;
+    private Integer points;
 }
